Added new realtime development task watch:ff.
Replaces the former gm + watch:gm tasks to aid in realtime development and debugging.

The new watch:ff task builds a new Firefox extension and autoloads it into Firefox in the most performant way possible.

Note that the stated performance goal is not accomplished yet; the current task averages at 2.8 secs (on Windows).