Inject script to let dreditor and page communicate
The content script needs to know when certain things happen so it can do some
work. One example is when Drupal changes the content of the page via AJAX, we
want to rerun the plugins. Communication is only allowed between content scripts
and the page via the window.postMessage() API.

A small communication handler is used to abstract the sending/listening of data
to/from the page. A small script is injected into the page that can figure out
when the page has done something of interest and then tell the content script by
sending a message (e.g., when Drupal.attachBehaviors() gets called, the content
script will get notified).